backend/requirements.txt
# Django
Django>=5.0,<6.0
django-ninja>=1.3,<2.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
PyJWT>=2.8,<3.0
django-ninja-jwt>=5.3,<6.0

# Utils
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
whitenoise>=6.6,<7.0

# Development
django-debug-toolbar>=4.2,<5.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-asyncio>=0.23,<1.0
